requirements.txt
discord.py==2.1.0
Flask==2.2.2
requests==2.28.2
PyYAML==6.0